SHTnsSpheres

Documentation for SHTnsSpheres.

SHTnsSpheres.InOut

SHTnsSpheres.erase

SHTnsSpheres.readable

SHTnsSpheres.shtns_use_threads

SHTnsSpheres.writable

SHTnsSpheres.InOut — Type

Wraps data to mark it as writeable.

SHTnsSpheres.erase — Method

y = erase(x)

Some functions need to make copies of input arguments to avoid modifying their contents and remain pure. Passing erase(x) as input argument is equivalent to passing x, except that it explicitly allows to modify the contents of x, thus avoiding copying and allocating.

SHTnsSpheres.readable — Method

Unwrap input argument. Used internally when we can promise that x will not be modified.

SHTnsSpheres.shtns_use_threads — Function

num_threads = shtns_use_threads(num_threads)

Call before any initialization of shtns to use mutliple threads. Returns the actual number of threads. If numthreads > 0, specifies the maximum number of threads that should be used. If numthreads <= 0, maximum number of threads is automatically set to the number of processors. If num_threads == 1, openmp will be disabled.

SHTnsSpheres.writable — Method

Unwrap input argument. Used internally when we cannot promise that x will not be modified.